Fix SDRAM 16->32 bit conversion
Due to incoherent component specifications I missed that no
conversion between the 16 bit sdram and 32 bit avalon bus
was happening.  This adds an avalon convertor into the mix.

Note that the avalon address busses are also not the correct
widths (though it still works ok) which should also be fixed.

Once the avalon slave/master interfaces are functorised over
the address / data widths this problem should not reoccur.